// Backend/src/enums/complaint-against.ts
export enum ComplaintAgainst {
    ACCOUNTANT_GENERAL_SINDH = 'Accountant General Sindh',
    AGRICULTURE_SUPPLY_AND_PRICES_DEPARTMENT = 'Agriculture Supply and Prices Department',
    AUQAF_RELIGIOUS_AFFAIRS_ZAKAT_AND_USHR_DEPARTMENT = 'Auqaf Religious Affairs Zakat and Ushr Department',
    BENAZIR_BHUTTO_SHAHEED_HUMAN_RESOURCE_RESEARCH_AND_DEVELOPMENT_BOARD = 'Benazir Bhutto Shaheed Human Resource Research and Development Board',
    COOPERATION_DEPARTMENT = 'Cooperation Department',
    CULTURE_TOURISM_AND_ANTIQUITIES_DEPARTMENT = 'Culture Tourism and Antiquities Department',
    EDUCATION_AND_LITERACY_DEPARTMENT_COLLEGES_EDUCATION = 'Education and Literacy Department (Colleges Education)',
    EDUCATION_AND_LITERACY_DEPARTMENT_SCHOOLS_EDUCATION = 'Education and Literacy Department (Schools Education)',
    ENERGY_DEPARTMENT = 'Energy Department',
    ENQUIRIES_AND_ANTI_CORRUPTION_ESTABLISHMENT = 'Enquiries and Anti Corruption Establishment',
    ENVIRONMENT_CLIMATE_CHANGE_AND_COSTAL_DEVELOPMENT_DEPARTMENT = 'Environment Climate Change and Costal Development Department',
    ENVIRONMENTAL_ALTERNATE_ENERGY = 'Environmental Alternate Energy',
    EXCISE_TAXATION_AND_NARCOTICS_DEPARTMENT = 'Excise Taxation and Narcotics Department',
    FINANCE_DEPARTMENT = 'Finance Department',
    FOREST_AND_WILDLIFE_DEPARTMENT = 'Forest and Wildlife Department',
    FOOD_DEPARTMENT = 'Food Department',
    HEALTH_DEPARTMENT = 'Health Department',
    HOME_DEPARTMENT = 'Home Department',
    HUMAN_RIGHTS_DEPARTMENT = 'Human Rights Department',
    INDUSTRIES_AND_COMMERCE_DEPARTMENT = 'Industries and Commerce Department',
    INFORMATION_AND_ARCHIVES_DEPARTMENT = 'Information and Archives Department',
    INFORMATION_SCIENCE_AND_TECHNOLOGY_DEPARTMENT = 'Information Science and Technology Department',
    IRRIGATION_DEPARTMENT = 'Irrigation Department',
    KARACHI_MUNICIPAL_CORPORATION = 'Karachi Municipal Corporation',
    KARACHI_DEVELOPMENT_AUTHORITY = 'Karachi Development Authority',
    KARACHI_WATER_AND_SEWERAGE_BOARD = 'Karachi Water and Sewerage Board',
    KATCHI_ABADIES_AND_SPATIAL_DEVELOPMENT = 'Katchi Abadies and Spatial Development',
    LABOUR_AND_HUMAN_RESOURCES_DEPARTMENT = 'Labour and Human Resources Department',
    LAW_AND_PARLIAMENTARY_AFFAIRS_DEPARTMENT = 'Law and Parliamentary Affairs Department',
    LIVESTOCK_AND_FISHERIES_DEPARTMENT = 'Live Stock and Fisheries Department',
    LOCAL_GOVERNMENT_AND_HOUSING_TOWN_PLANNING_DEPARTMENT = 'Local Government and Housing Town Planning Department',
    LYARI_DEVELOPMENT_AUTHORITY = 'Lyari Development Authority',
    MALI_DEVELOPMENT_AUTHORITY = 'Mali Development Authority',
    MINES_AND_MINERAL_DEVELOPMENT_DEPARTMENT = 'Mines and Mineral Development Department',
    MINORITIES_AFFAIRS_DEPARTMENT = 'Minorities Affairs Department',
    PLANNING_AND_DEVELOPMENT_DEPARTMENT = 'Planning and Development Department',
    POLICE_DEPARTMENT = 'Police Department',
    POPULATION_AND_WELFARE_DEPARTMENT = 'Population and Welfare Department',
    POWER_DEPARTMENT = 'Power Department',
    PUBLIC_HEALTH_ENGINEERING_AND_RURAL_DEVELOPMENT_DEPARTMENT = 'Public Health Engineering and Rural Development Department',
    REVENUE_DEPARTMENT = 'Revenue Department',
    SERVICES_GENERAL_ADMINISTRATION_AND_COORDINATION_DEPARTMENT = 'Services General Administration and Coordination Department',
    SINDH_BUILDING_CONTROL_AUTHORITY = 'Sindh Building Control Authority',
    SINDH_EMPLOYEE_SOCIAL_SECURITY_INSTITUTION = 'Sindh Employees Social Security Institution',
    SINDH_HIGHER_EDUCATION_COMMISSION_KARACHI = 'Sindh Higher Education Commission Karachi',
    SINDH_PUBLIC_SERVICE_COMMISSION = 'Sindh Public Service Commission',
    SOCIAL_WELFARE_DEPARTMENT = 'Social Welfare Department',
    SPECIAL_EDUCATION_DEPARTMENT = 'Special Education Department',
    SPORT_AND_YOUTH_AFFAIRS_DEPARTMENT = 'Sport and Youth Affairs Department',
    TRANSPORT_AND_MASS_TRANSIT_DEPARTMENT = 'Transport and Mass Transit Department',
    WOMEN_DEVELOPMENT_DEPARTMENT = 'Women Development Department',
    WORK_AND_SERVICES_DEPARTMENT = 'Work and Services Department',
}