// Backend/src/enums/complaint-office.ts
export enum ComplaintOffice {
    KARACHI_CENTRAL = 'Karachi (Central)',
    KARACHI_EAST = 'Karachi (East)',
    KARACHI_SOUTH = 'Karachi (South)',
    THATTA = 'Thatta',
    SUKKUR = 'Sukkur',
    HYDERABAD = 'Hyderabad',
    BADIN = 'Badin',
    NAUSHAHRO_FEROZE = 'Naushahro Feroze',
    LARKANA = 'Larkana',
    MIRPURKHAS = 'Mirpurkhas',
    MITHI = 'Mithi',
    JACOBABAD = 'Jacobabad',
    DADU = 'Dadu',
    NAWABSHAH = 'Nawabshah',
    KHAIRPUR = 'Khairpur',
    GHOTKI = 'Ghotki'
}